src/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Django==4.2.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.6.3
Werkzeug==2.3.4
WTForms==3.0.1
